feat: store equals for both content and equals multihash
We want to return equals claims when quired with either the content cid or the equals cid, so store 2 records in dynamo for equals claims.

`assert/equals` claims provide 2 CIDs. The `content` CID and the `equals` CID that we claim the `content` CID is equivelant to. (e.g. different hash fn for same bytes)

The dynamo records are `{ content: "base58btc-multihash", claim: CID }`. The actual bytes of the claim are stored in s3, only once per claim.

For equals claims we store an additional record that sets the `content` to the multihash of the provided `equals` CID.

**Alternatively** we could have stored an `equals` property on dynamo records for equals claims, and added a global secondary index for that field, and updated the query to check both the content field and the equals field (and the secondary index) for every query, but this would incur additonal read costs for *every* query. It is assumed that storing a extra tuple of [string,cid] per equals claim will be cheap, and avoids a larger refactor of the claims db.

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>